Introduce common base class for Flink systems
